Simple Wifi fix on Ubuntu 17.04
There's a bug in this version of network-manager (1.4.4) that shipped with the initial release of 17.04. It randomizes the MAC address of a wifi chipset when connecting to a new network for security purposes but the end result is that some wifi devices won't connect or won't stay connected to the network with this feature enabled.

The good news is this is actually really easy to fix! To disable the randomized mac address switching and fix your wifi, you just have to edit a config file.